+++ packages/access-control/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @creezio/access-control
 
+## 0.22.0
+
+### Patch Changes
+
+- @creezio/shell-ui@0.22.0
+- @creezio/auth@0.22.0
+
+## 0.21.0
+
+### Patch Changes
+
+- @creezio/shell-ui@0.21.0
+- @creezio/auth@0.21.0
+
 ## 0.20.0
 
 ### Patch Changes