+++ packages/database/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @creezio/database
 
+## 0.22.0
+
+### Patch Changes
+
+- @creezio/platform-core@0.22.0
+
+## 0.21.0
+
+### Patch Changes
+
+- @creezio/platform-core@0.21.0
+
 ## 0.20.0
 
 ### Patch Changes