+++ packages/product-hub/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @creezio/product-hub
 
+## 0.22.0
+
+### Patch Changes
+
+- @creezio/brand-config@0.22.0
+- @creezio/platform-core@0.22.0
+
+## 0.21.0
+
+### Patch Changes
+
+- @creezio/brand-config@0.21.0
+- @creezio/platform-core@0.21.0
+
 ## 0.20.0
 
 ### Patch Changes